+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v3.1.0
-- **[feat]** - Add`BottomFixedPaper` component
+## v3.1.1
+- **[fix]** -`BottomFixedPaper` props position
+- **[fix]** -`BottomFixedPaper` padding props